refactor: replace custom tool call validation with `vi.spyOn` in HBAR transfer tests
- Removed outdated `getToolCallFromResult` and `expectToolCall` functions from `langchain-test-setup.ts`.
- Refactored integration tests to use `vi.spyOn` for validating tool interactions directly with `getHederaAgentKitAPI`.
- Enhanced test setup to ensure consistent mocking and validation of API calls.
- Improved readability and maintainability of HBAR transfer tests by using `coreAccountPluginToolNames`.
